Refactor database package and add SportsbookPreference
