+++ src/database/__init__.py
@@ -1,0 +1,12 @@
+from .models import Base, User, SportsbookPreference
+from .session import engine, SessionLocal, init_db, get_db_session
+
+__all__ = [
+    "Base",
+    "User",
+    "SportsbookPreference",
+    "engine",
+    "SessionLocal",
+    "init_db",
+    "get_db_session",
+]
